# docs/OpenAPI.yaml
openapi: 3.0.3
info:
  title: DCS Server Management API
  version: 1.0.0
  description: >
    Standard REST API for managing DCS Dedicated Server hosting.
    Designed for an Orchestrator (public edge) + Host Agents (internal nodes).
    Long-running actions are modeled as Jobs. Real-time notifications via SSE/WebSocket.
servers:
  - url: https://orchestrator.example.com/api/v1
    description: Orchestrator API (public edge)
  - url: http://agent.local:8787/agent/v1
    description: Agent API (internal node)
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /instances/{instanceId}:
    get:
      tags: [Instances]
      summary: Get instance definition + latest state
      operationId: getInstance
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Instance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "404":
          $ref: "#/components/responses/Problem"

    patch:
      tags: [Instances]
      summary: Patch instance definition
      operationId: updateInstance
      parameters:
        - $ref: "#/components/parameters/instanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstancePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"

    delete:
      tags: [Instances]
      summary: Delete instance definition
      operationId: deleteInstance
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/Problem"
...
  /instances/{instanceId}/missions:
    get:
      tags: [Instances, Missions]
      summary: List missions available to instance
      operationId: listMissions
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Missions
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Mission" }
...
  /instances/{instanceId}/status:
    get:
      tags: [Agent]
      summary: Local runtime status
      operationId: agentGetInstanceStatus
      servers:
        - url: http://agent.local:8787/agent/v1
      parameters:
        - $ref: "#/components/parameters/instanceId"
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstanceRuntime"
  /instances/{instanceId}/actions/{action}:
    post:
      tags: [Agent, Jobs]
      summary: Execute a local action (async)
      operationId: agentAction
      servers:
        - url: http://agent.local:8787/agent/v1
      parameters:
        - $ref: "#/components/parameters/instanceId"
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [start, stop, restart, update, mission_load, logs_bundle]
        - $ref: "#/components/parameters/idempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "202":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobAccepted"
        "400":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Optional idempotency key to avoid duplicate actions.
...
    instanceId:
      name: instanceId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Problem:
      description: Error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, example: "https://errors.example.com/instance-not-found" }
        title: { type: string, example: "Instance not found" }
        status: { type: integer, example: 404 }
        detail: { type: string }
        instance: { type: string, description: "Request correlation id / trace id" }
...
    InstanceStatus:
      type: string
      enum: [unknown, stopped, starting, running, stopping, updating, error]
...
    InstancePatch:
      type: object
      properties:
        name: { type: string }
        tags:
          type: array
          items: { type: string }
        ports:
          $ref: "#/components/schemas/InstancePorts"
        service:
          $ref: "#/components/schemas/ServiceBinding"
        config:
          $ref: "#/components/schemas/InstanceConfigPatch"
    Instance:
      type: object
      required: [id, serverId, hostId, name, paths, ports, service, createdAt]
      properties:
        id: { type: string, example: "inst_01HXYZ" }
        serverId: { type: string }
        hostId: { type: string }
        name: { type: string }
        tags:
          type: array
          items: { type: string }
        paths:
          $ref: "#/components/schemas/InstancePaths"
        ports:
          $ref: "#/components/schemas/InstancePorts"
        service:
          $ref: "#/components/schemas/ServiceBinding"
        config:
          $ref: "#/components/schemas/InstanceConfig"
        runtime:
          $ref: "#/components/schemas/InstanceRuntime"
        createdAt: { type: string, format: date-time }
...
    InstancePaths:
      type: object
      required: [dcsRoot, savedGames, missionDir, logsDir]
      properties:
        dcsRoot: { type: string, example: "D:\\DCS\\DedicatedServer" }
        savedGames: { type: string, example: "C:\\Users\\User\\Saved Games\\DCS.server1" }
        missionDir: { type: string, example: "C:\\Missions\\server1" }
        logsDir: { type: string, example: "C:\\Users\\User\\Saved Games\\DCS.server1\\Logs" }
    InstancePorts:
      type: object
      required: [gamePort, webGuiPort]
      properties:
        gamePort: { type: integer, minimum: 1, maximum: 65535, example: 10308 }
        webGuiPort: { type: integer, minimum: 1, maximum: 65535, example: 8088 }
        srsPort: { type: integer, minimum: 1, maximum: 65535, nullable: true }
        tacviewPort: { type: integer, minimum: 1, maximum: 65535, nullable: true }
    ServiceBinding:
      type: object
      required: [type, name]
      properties:
        type:
          type: string
          enum: [nssm, systemd, docker]
          description: Service manager type for this instance
        name:
          type: string
          description: Service name (e.g., NSSM service name)
        autoStart:
          type: boolean
          default: true
    InstanceRuntime:
      type: object
      required: [status, observedAt]
      properties:
        status:
          $ref: "#/components/schemas/InstanceStatus"
        observedAt: { type: string, format: date-time }
        pid: { type: integer, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        uptimeSeconds: { type: integer, nullable: true }
        lastExitCode: { type: integer, nullable: true }
        lastError: { type: string, nullable: true }
...
    Mission:
      type: object
      required: [id, name, path]
      properties:
        id: { type: string, example: "miz_01HXYZ" }
        name: { type: string, example: "Goonfront Caucasus v0.3" }
        path: { type: string, example: "C:\\Missions\\server1\\goonfront_v0_3.miz" }
        sizeBytes: { type: integer, format: int64, nullable: true }
        modifiedAt: { type: string, format: date-time, nullable: true }
...
    InstanceConfig:
      type: object
      description: Safe subset of configuration (avoid secrets / unsafe knobs).
      properties:
        maxPlayers: { type: integer, minimum: 1, maximum: 200, nullable: true }
        passwordRequired: { type: boolean, nullable: true }
        missionRotation:
          type: object
          properties:
            enabled: { type: boolean, default: false }
            intervalMinutes: { type: integer, minimum: 5, maximum: 1440, nullable: true }
        logSettings:
          type: object
          properties:
            errorScanEnabled: { type: boolean, default: true }
            errorScanIntervalSeconds: { type: integer, minimum: 10, maximum: 600, default: 60 }
    InstanceConfigPatch:
      type: object
      properties:
        maxPlayers: { type: integer, minimum: 1, maximum: 200, nullable: true }
        passwordRequired: { type: boolean, nullable: true }
        missionRotation:
          type: object
          properties:
            enabled: { type: boolean }
            intervalMinutes: { type: integer, minimum: 5, maximum: 1440, nullable: true }
        logSettings:
          type: object
          properties:
            errorScanEnabled: { type: boolean }
            errorScanIntervalSeconds: { type: integer, minimum: 10, maximum: 600 }
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled]
...
    JobAccepted:
      type: object
      required: [jobId, status]
      properties:
        jobId: { type: string, example: "job_01HXYZ" }
        status:
          $ref: "#/components/schemas/JobStatus"
...